use textEditor slice
