+++ packages/compass-aggregations/src/modules/pipeline-builder/text-editor.ts
@@ -1,5 +1,5 @@
 import type { Reducer } from 'redux';
-import type { Document, MongoServerError } from 'mongodb';
+import type { AggregateOptions, Document, MongoServerError } from 'mongodb';
 import type { PipelineBuilderThunkAction } from '..';
 import { DEFAULT_MAX_TIME_MS } from '../../constants';
 import type { PreviewOptions } from './pipeline-preview-manager';
@@ -15,15 +15,21 @@ import type { PipelineModeToggledAction } from './pipeline-mode';
 import { getStageOperator } from '../../utils/stage';
 import { CONFIRM_NEW, NEW_PIPELINE } from '../import-pipeline';
 import { RESTORE_PIPELINE } from '../saved-pipeline';
+import { aggregatePipeline } from '../../utils/cancellable-aggregation';
+import { gotoOutResults } from '../out-results-fn';
+import { globalAppRegistryEmit } from '@mongodb-js/mongodb-redux-common/app-registry';
 
 export const enum EditorActionTypes {
   EditorPreviewFetch = 'compass-aggregations/pipeline-builder/text-editor/TextEditorPreviewFetch',
   EditorPreviewFetchSuccess = 'compass-aggregations/pipeline-builder/text-editor/TextEditorPreviewFetchSuccess',
   EditorPreviewFetchError = 'compass-aggregations/pipeline-builder/text-editor/TextEditorPreviewFetchError',
   EditorValueChange = 'compass-aggregations/pipeline-builder/text-editor/TextEditorValueChange',
+  OutputStageFetchStarted = 'compass-aggregations/pipeline-builder/text-editor/OutputStageFetchStarted',
+  OutputStageFetchSucceded = 'compass-aggregations/pipeline-builder/text-editor/OutputStageFetchSucceded',
+  OutputStageFetchFailed = 'compass-aggregations/pipeline-builder/text-editor/OutputStageFetchFailed',
 };
 
-export type EditorValueChangeAction = {
+type EditorValueChangeAction = {
   type: EditorActionTypes.EditorValueChange;
   pipelineText: string;
   pipeline: Document[] | null;
@@ -44,6 +50,19 @@ type EditorPreviewFetchErrorAction = {
   serverError: MongoServerError;
 };
 
+type OutputStageFetchStartedAction = {
+  type: EditorActionTypes.OutputStageFetchStarted;
+};
+
+type OutputStageFetchSuccededAction = {
+  type: EditorActionTypes.OutputStageFetchSucceded;
+};
+
+type OutputStageFetchFailedAction = {
+  type: EditorActionTypes.OutputStageFetchFailed;
+  serverError: MongoServerError;
+};
+
 export type TextEditorState = {
   pipelineText: string;
   stageOperators: string[];
@@ -51,6 +70,10 @@ export type TextEditorState = {
   serverError: MongoServerError | null;
   loading: boolean;
   previewDocs: Document[] | null;
+  outputStage: {
+    isLoading: boolean;
+    isComplete: boolean;
+  }
 };
 
 const INITIAL_STATE: TextEditorState = {
@@ -60,6 +83,10 @@ const INITIAL_STATE: TextEditorState = {
   serverError: null,
   loading: false,
   previewDocs: null,
+  outputStage: {
+    isLoading: false,
+    isComplete: false,
+  }
 };
 
 const reducer: Reducer<TextEditorState> = (state = INITIAL_STATE, action) => {
@@ -81,6 +108,10 @@ const reducer: Reducer<TextEditorState> = (state = INITIAL_STATE, action) => {
       : state.stageOperators;
     return {
       ...state,
+      outputStage:{
+        isLoading: false,
+        isComplete: false,
+      },
       serverError: null,
       previewDocs: null,
       pipelineText: action.pipelineText,
@@ -133,10 +164,58 @@ const reducer: Reducer<TextEditorState> = (state = INITIAL_STATE, action) => {
     };
   }
 
+  if (
+    isAction<OutputStageFetchStartedAction>(
+      action,
+      EditorActionTypes.OutputStageFetchStarted
+    )
+  ) {
+    return {
+      ...state,
+      serverError: null,
+      outputStage: {
+        isLoading: true,
+        isComplete: false,
+      },
+    };
+  }
+
+  if (
+    isAction<OutputStageFetchSuccededAction>(
+      action,
+      EditorActionTypes.OutputStageFetchSucceded
+    )
+  ) {
+    return {
+      ...state,
+      serverError: null,
+      outputStage: {
+        isLoading: false,
+        isComplete: true,
+      },
+    };
+  }
+
+  if (
+    isAction<OutputStageFetchFailedAction>(
+      action,
+      EditorActionTypes.OutputStageFetchFailed
+    )
+  ) {
+    return {
+      ...state,
+      serverError: action.serverError,
+      outputStage: {
+        isLoading: false,
+        isComplete: false,
+      },
+    };
+  }
+
   return state;
 };
 
-export function canRunPipeline(
+function canRunPipeline(
   autoPreview: boolean,
   syntaxErrors: PipelineParserError[],
 ) {
@@ -214,5 +293,70 @@ export const changeEditorValue = (
     void dispatch(loadPreviewForPipeline());
   };
 };
+
+export const runPipelineWithOutputStage = (
+  ): PipelineBuilderThunkAction<Promise<void>> => {
+    return async (dispatch, getState, { pipelineBuilder }) => {
+      const {
+        autoPreview,
+        isAtlasDeployed,
+        dataService: { dataService },
+        namespace,
+        maxTimeMS,
+        collationString,
+      } = getState();
+  
+  
+      if (!dataService || !isAtlasDeployed) {
+        return;
+      }
+  
+      if (!canRunPipeline(autoPreview, pipelineBuilder.syntaxError)) {
+        return;
+      }
+  
+      try {
+        dispatch({ type: EditorActionTypes.OutputStageFetchStarted });
+        const pipeline = pipelineBuilder.getPipelineFromSource();
+        const options: AggregateOptions = {
+          maxTimeMS: maxTimeMS ?? DEFAULT_MAX_TIME_MS,
+          collation: collationString.value ?? undefined
+        };
+        const { signal } = new AbortController();
+        await aggregatePipeline({
+          dataService,
+          signal,
+          namespace,
+          pipeline,
+          options
+        });
+        dispatch({
+          type: EditorActionTypes.OutputStageFetchSucceded,
+        });
+        dispatch(globalAppRegistryEmit('agg-pipeline-out-executed'));
+      } catch (error) {
+        dispatch({
+          type: EditorActionTypes.OutputStageFetchFailed,
+          serverError: error,
+        });
+      }
+    };
+  };
+  
+  export const gotoOutputStageCollection = (
+  ): PipelineBuilderThunkAction<void> => {
+    return (dispatch, getState) => {
+      const {
+        pipelineBuilder: {
+          textEditor: {
+            stageOperators
+          }
+        }
+      } = getState();
+      // $out or $merge is always last stage
+      const lastStageIndex = stageOperators.length - 1;
+      dispatch(gotoOutResults(lastStageIndex));
+    };
+  };
 
 export default reducer;
